pyproject.toml
[project]
name = "agnoagent"
version = "0.1.0"
description = "Multi-Agent System with agno, a2a-sdk and MCP integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.16",
    "agno>=1.7.6",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "trafilatura>=1.8.0",
    "ntplib>=0.4.0",
    "python-dotenv>=1.0.0",
    "openai>=1.97.1",
    "ddgs>=9.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
